+++ debian/changelog
@@ -1,3 +1,9 @@
+wb-mqtt-homeui (2.222.0) stable; urgency=medium
+
+  * Add page name in title
+
+ -- Victor Vedenin <victor.vedenin@wirenboard.com>  Fri, 05 Jun 2026 14:56:00 +0300
+
 wb-mqtt-homeui (2.221.1) stable; urgency=medium
 
   * Fix endless loader on device scan add
