+++ frontend/src/i18n/locales/ru.json
@@ -140,7 +140,7 @@
       }
    },
    "scan": {
-      "title": "Новые устройства",
+      "title": "Поиск новых устройств",
       "search-disconnected-title": "Поиск неотвечающего устройства",
       "errors": {
          "scan": "Не удалось запустить сканирование",
@@ -960,6 +960,7 @@
          "add-prefix-to-topic": "Добавлять префикс (/client/user_id/)",
          "common-settings": "Общие настройки",
          "show-system-devices": "Показывать системные устройства",
+         "show-page-in-title": "Показывать имя страницы в заголовке вкладки браузера",
          "show-widgets-page": "Отображать в навигации страницу Виджеты",
          "language": "Язык интерфейса",
          "default-dashboard": "Панель по-умолчанию",
