+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/capabilities/index.ts
@@ -1,0 +1,15 @@
+export {
+  getAvailableModeInstances,
+  getAvailableRangeInstances,
+  getAvailableColorModels,
+  getAvailableToggleInstances,
+  getColorModelLabel,
+  getCurrentColorModel,
+  RANGE_LIMITS_DEFAULT,
+  RANGE_LIMITS_LOCKED
+} from './helpers';
+export { RangeCapability } from './range';
+export { ColorSettingCapability } from './color-setting';
+export { ModeCapability } from './mode';
+export { OnOffCapability } from './on-off';
+export { ToggleCapability } from './toggle';
