+++ frontend/src/pages/integrations/alice/components/smart-device/components/device-skills/styles.css
@@ -4,6 +4,11 @@
     border-radius: var(--card-border-radius);
 }
 
+.aliceDeviceSkills-title {
+    margin: 10px 0;
+    font-size: var(--font-size-medium);
+}
+
 .aliceDeviceSkills-grid {
     display: grid;
     grid-template-columns: minmax(100px, 150px) minmax(100px, 1fr) minmax(100px, 150px) minmax(auto, 1fr) 22px;
